pyproject.toml
#[project]
#name = "seefood-py"
#version = "0.1.0"
#description = "Add your description here"
#readme = "README.md"
#requires-python = ">=3.11"
#dependencies = [
#    "opencv-python>=4.12.0.88",
#    "setuptools>=80.9.0",
#    "smbus2>=0.5.0",
#    "sparkfun-qwiic-nau7802>=2.0.0",
#    "vl53l0x",
#]
#
#[tool.uv.sources]
#vl53l0x = { git = "https://github.com/pimoroni/VL53L0X-python.git" }

[project]
name = "seefood-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
# Explicitly pin to Python 3.11 so uv uses the stable ABI
requires-python = "==3.11.*"
dependencies = [
    "asyncio>=4.0.0",
    "opencv-python>=4.12.0.88",
    "phidget22>=1.23.20250925",
    "setuptools>=80.9.0",
    "smbus2>=0.5.0",
    "sparkfun-qwiic-nau7802>=2.0.0",
    "vl53l0x",
]

[tool.uv.sources]
# Pull from GitHub instead of PyPI wheel (builds from source)
vl53l0x = { git = "https://github.com/Gadgetoid/VL53L0X-python.git" }
